requirements.txt
flask
flask-cors
piper-tts
gunicorn